requirements-test.txt
pytest
psycopg2
solrpy
